+++ src/features/laporan/aging-premi/actions/useLapAgingPremiDownload.ts
@@ -1,88 +1,24 @@
-import { useState } from 'react';
 import { LaporanAgingPremiRow } from "@/lib/laporan/laporan-aging-premi/types";
 import { generateAgingPremiPDF } from "@/features/laporan/aging-premi/actions/pdf-aging-premi-generator";
-import { formatDate } from "@/lib/utils/formatDate";
 import { generateAgingPremiExcel } from '@/features/laporan/aging-premi/actions/aging-premi-excel-generator';
-import triggerDownload from '@/lib/utils/triggerDownload';
+import useLaporanExporter from '../../useLaporanExporter';
 
 interface ExporterProps {
   rowData: LaporanAgingPremiRow[];
   startDate: string;
   endDate: string;
 }
-
 export default function useLapAgingPremiDownload({
   rowData,
   startDate,
   endDate,
 }: ExporterProps) {
-  const [pdfPreview, setPdfPreview] = useState({ isOpen: false, dataUrl: "" });
-  const [isExcelPreviewOpen, setIsExcelPreviewOpen] = useState(false);
-
-  const handlePDFPreview = () => {
-    if (rowData.length === 0) {
-      console.log("No data to generate PDF.");
-      // In a real app, you might use a toast notification here.
-      return;
-    }
-    const dataUrl = generateAgingPremiPDF(rowData, startDate, endDate);
-    setPdfPreview({ isOpen: true, dataUrl });
-  };
-
-  const handlePDFDownload = () => {
-    const datePart =
-      startDate && endDate
-        ? `${formatDate(startDate)}_to_${formatDate(endDate)}`
-        : "all_time";
-    const fileName = `Laporan_Aging_Premi_${datePart}.pdf`.replace(/ /g, '_');
-    triggerDownload(pdfPreview.dataUrl, fileName);
-  };
-
-  const closePDFPreview = () => {
-    setPdfPreview({ isOpen: false, dataUrl: "" });
-  };
-
-  // =================================================================================
-  //                                  Excel Generation
-  // =================================================================================
-  const handleExcelPreview = () => {
-    if (rowData.length === 0) {
-      console.log("No data to preview.");
-      return;
-    }
-    setIsExcelPreviewOpen(true);
-  };
-
-  const closeExcelPreview = () => {
-    setIsExcelPreviewOpen(false);
-  };
-
-  const handleExcelExport = async () => {
-    if (rowData.length === 0) {
-      console.log("No data to generate Excel file.");
-      return;
-    }
-
-    const excelBlob = await generateAgingPremiExcel(rowData, startDate, endDate);
-    const dataUrl = URL.createObjectURL(excelBlob);
-    const datePart =
-      startDate && endDate
-        ? `${formatDate(startDate)}_to_${formatDate(endDate)}`
-        : "all_time";
-    const fileName = `Laporan_Aging_Premi_${datePart}.xlsx`;
-
-    triggerDownload(dataUrl, fileName);
-    closeExcelPreview(); // Close dialog after download starts
-  };
-
-  return {
-    pdfPreview,
-    handlePDFPreview,
-    handlePDFDownload,
-    closePDFPreview,
-    isExcelPreviewOpen,
-    handleExcelPreview,
-    closeExcelPreview,
-    handleExcelExport,
-  };
+  return useLaporanExporter<LaporanAgingPremiRow>({
+    rowData,
+    startDate,
+    endDate,
+    fileNamePrefix: "Laporan_Aging_Premi",
+    pdfGenerator: generateAgingPremiPDF,
+    excelGenerator: generateAgingPremiExcel,
+  });
 }
